pyproject.toml
[tool.poetry]
name = "eval-llm"
version = "0.1.0"
description = "An LLM evaluation tool for running datasets and generating reports."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://your-project-homepage.com"
repository = "https://github.com/yourusername/eval-llm"
keywords = ["LLM", "evaluation", "machine learning", "NLP"]

# Specify the package directory
packages = [
    { include = "eval_llm", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.28.2"
PyYAML = "^6.0"
jinja2 = "^3.1.2"

[tool.poetry.scripts]
eval-llm = "eval_llm.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
